requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
httpx>=0.26.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
cryptography>=42.0.0
pyjwt>=2.8.0
structlog>=24.1.0
slowapi>=0.1.9
